+++ src/main/scala/info/gratour/jt808common/codec/decoder/impl/MBDecoder808_0201_QryLocationAck.scala
@@ -16,7 +16,15 @@ import io.netty.buffer.ByteBuf
 object MBDecoder808_0201_QryLocationAck extends JT808MsgBodyDecoder[JT808Msg_0201_QryLocationAck]{
   override def decodeMsgBody(protoVer: Byte, adasDialect: AdasDialect, m: JT808Msg_0201_QryLocationAck, body: ByteBuf, tempBuf: Array[Byte]): Unit = {
     m.setAckSeqNo(body.readUnsignedShort())
-    val t = MBDecoder808_Track.decodeTrack(protoVer, m, body, tempBuf, System.currentTimeMillis(), retransmit = false)
+    val t = MBDecoder808_Track.decodeTrack(
+      protoVer,
+      adasDialect,
+      m,
+      body,
+      tempBuf,
+      System.currentTimeMillis(),
+      retransmit = false
+    )
     val ackParams = new JT808AckParams_0201_QryLocationAck
     ackParams.setTrk(t)
     m.setAckParams(ackParams)