+++ src/main/scala/info/gratour/jt808common/protocol/JTAdasAlarmType.java
@@ -1,0 +1,6 @@
+package info.gratour.jt808common.protocol;
+
+public enum JTAdasAlarmType {
+
+
+}