+++ src/main/scala/info/gratour/jt808common/protocol/msg/types/trk/AdasOverHeightAlmAddt.java
@@ -1,36 +1,25 @@
-/*******************************************************************************
- *  Copyright (c) 2019, 2022 lucendar.com.
- *  All rights reserved.
- *
- *  Contributors:
- *     KwanKin Yau (alphax@vip.163.com) - initial API and implementation
- *******************************************************************************/
 package info.gratour.jt808common.protocol.msg.types.trk;
 
 import java.util.StringJoiner;
 
-public class AdasAiRecogAlmAddt implements Cloneable {
+public class AdasOverHeightAlmAddt implements Cloneable {
 
-    public static final byte FLAG__NOT_AVAILABLE = 0;
     public static final byte FLAG__START = 1;
     public static final byte FLAG__END = 2;
+    public static final byte FLAG__CONTINUOUSLY = 3;
 
     /**
-     * 超远报警
+     * 超过限定高度报警
      */
-    public static final short TYP__OVERCROWDING = 1;
-
-    /**
-     * 未巡检乘客安全带（事件）
-     */
-    public static final short TYP__SEAT_BELT_NOT_CHECKED = 0x80;
+    public static final byte TYP__OVER_HEIGHT = 1;
 
     private int almId;
     private byte flag;
-    private short typ;
-    private byte lvl;
+    private byte typ;
+    private byte vehHeight;
+    private byte heightRestriction;
     private short spd;
-    private short alt;
+    private int alt;
     private double lat;
     private double lng;
     private long tm;
@@ -53,20 +42,28 @@ public class AdasAiRecogAlmAddt implements Cloneable {
         this.flag = flag;
     }
 
-    public short getTyp() {
+    public byte getTyp() {
         return typ;
     }
 
-    public void setTyp(short typ) {
+    public void setTyp(byte typ) {
         this.typ = typ;
     }
 
-    public byte getLvl() {
-        return lvl;
+    public byte getVehHeight() {
+        return vehHeight;
     }
 
-    public void setLvl(byte lvl) {
-        this.lvl = lvl;
+    public void setVehHeight(byte vehHeight) {
+        this.vehHeight = vehHeight;
+    }
+
+    public byte getHeightRestriction() {
+        return heightRestriction;
+    }
+
+    public void setHeightRestriction(byte heightRestriction) {
+        this.heightRestriction = heightRestriction;
     }
 
     public short getSpd() {
@@ -77,11 +74,11 @@ public class AdasAiRecogAlmAddt implements Cloneable {
         this.spd = spd;
     }
 
-    public short getAlt() {
+    public int getAlt() {
         return alt;
     }
 
-    public void setAlt(short alt) {
+    public void setAlt(int alt) {
         this.alt = alt;
     }
 
@@ -126,21 +123,22 @@ public class AdasAiRecogAlmAddt implements Cloneable {
     }
 
     @Override
-    public AdasAiRecogAlmAddt clone() {
+    public AdasOverHeightAlmAddt clone() {
         try {
-            return (AdasAiRecogAlmAddt) super.clone();
+            return (AdasOverHeightAlmAddt) super.clone();
         } catch (CloneNotSupportedException e) {
-            throw new AssertionError();
+            throw new RuntimeException(e);
         }
     }
 
     @Override
     public String toString() {
-        return new StringJoiner(", ", AdasAiRecogAlmAddt.class.getSimpleName() + "[", "]")
+        return new StringJoiner(", ", AdasOverHeightAlmAddt.class.getSimpleName() + "[", "]")
                 .add("almId=" + almId)
                 .add("flag=" + flag)
                 .add("typ=" + typ)
-                .add("lvl=" + lvl)
+                .add("vehHeight=" + vehHeight)
+                .add("heightRestriction=" + heightRestriction)
                 .add("spd=" + spd)
                 .add("alt=" + alt)
                 .add("lat=" + lat)